requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==20.2.0
certifi==2020.4.5.1
chardet==3.0.4
discord.py==1.4.1
idna==2.9
multidict==4.7.6
mysql-connector-python==8.0.20
praw==7.0.0
prawcore==1.3.0
protobuf==3.12.1
requests==2.23.0
schedule==0.6.0
six==1.15.0
update-checker==0.17
urllib3==1.25.9
websocket-client==0.57.0
yarl==1.5.1
